Decompose Model into sub-namespaces
