+++ tests/AnalyzeSpec.hs
@@ -40,11 +40,10 @@ import           Pact.Types.Runtime           (ModuleData, eeRefStore,
                                                rsModules)
 
 import           Pact.Analyze.Check
-import           Pact.Analyze.Model           (linearizedTrace)
+import qualified Pact.Analyze.Model           as Model
 import           Pact.Analyze.Parse           (PreProp (..), TableEnv,
                                                expToProp, inferProp)
 import           Pact.Analyze.PrenexNormalize (prenexConvert)
-import qualified Pact.Analyze.Model.Dot       as Dot
 import           Pact.Analyze.Types
 import           Pact.Analyze.Util            ((...))
 
@@ -104,7 +103,7 @@ renderTestFailure = \case
       Nothing -> pure ""
       Just m -> do
         let fp = "/tmp/execution-graph.dot"
-        Dot.render fp m
+        Model.renderDot fp m
         pure $ "\n\nrendered execution graph to DOT: " ++ fp
 
     pure $ T.unpack (describeCheckFailure cf) ++ svgInfo
@@ -1943,7 +1942,7 @@ spec = describe "analyze" $ do
           it "produces the correct trace" $
             case res of
               Just (TestCheckFailure (falsifyingModel -> Just model)) -> do
-                let trace = _etEvents (linearizedTrace model)
+                let trace = _etEvents (Model.linearize model)
                 unless (tests `match` trace) $ HUnit.assertFailure $
                   "trace doesn't match:\n\n" ++ show trace
               _ ->
